+++ fastapi_errors_plus/__init__.py
@@ -16,4 +16,4 @@ __all__ = [
     "BaseErrorDTO",
     "StandardErrorDTO",
 ]
-__version__ = "0.9.1"
+__version__ = "0.9.2"